[fix] minor refactoring
