+++ Sources/Entities/Encryption/AsyncSignerProtocol.swift
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2023 European Commission
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import Foundation
+@preconcurrency import JOSESwift
+
+public protocol AsyncSignerProtocol: Sendable {
+  func signAsync(_ header: Data, _ payload: Data) async throws -> Data
+}
+ 
+class PrecomputedSigner: JOSESwift.SignerProtocol {
+  var algorithm: JOSESwift.SignatureAlgorithm
+  let signature: Data
+  
+  init(signature: Data, algorithm: JOSESwift.SignatureAlgorithm) {
+    self.algorithm = algorithm
+    self.signature = signature
+  }
+  
+  func sign(_ signingInput: Data) throws -> Data {
+    return signature
+  }
+}
+
